c++: noexcept and pointer to member function type [PR113108]
We ICE in nothrow_spec_p because it got a DEFERRED_NOEXCEPT.
This DEFERRED_NOEXCEPT was created in implicitly_declare_fn
when declaring

  Foo& operator=(Foo&&) = default;

in the test.  The problem is that in resolve_overloaded_unification
we call maybe_instantiate_noexcept before try_one_overload only in
the TEMPLATE_ID_EXPR case.

	PR c++/113108

gcc/cp/ChangeLog:

	* pt.cc (resolve_overloaded_unification): Call
	maybe_instantiate_noexcept.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp1z/noexcept-type28.C: New test.