aarch64: Use is_attribute_namespace_p and get_attribute_name inside aarch64_lookup_shared_state_flags [PR116598]
The code in aarch64_lookup_shared_state_flags all C++11 attributes on the function type
had a namespace associated with them. But with the addition of reproducible/unsequenced,
this is not true.

This fixes the issue by using is_attribute_namespace_p instead of manually figuring out
the namespace is named "arm" and uses get_attribute_name instead of manually grabbing
the attribute name.

Built and tested for aarch64-linux-gnu.

gcc/ChangeLog:

	PR target/116598
	* config/aarch64/aarch64.cc (aarch64_lookup_shared_state_flags): Use
	is_attribute_namespace_p and get_attribute_name instead of manually grabbing
	the namespace and name of the attribute.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>